+++ source/app/service-providers/assets/get-config-template.ts
@@ -209,6 +209,7 @@ export default function getConfigTemplate (): ConfigOptions {
       // By default, Windows and macOS systems use the system accent color instead of the theme
       useSystemAccentColor: [ 'darwin', 'win32' ].includes(process.platform),
       hideToolbarInDistractionFree: false,
+      markdownFileExtensions: false,
       imageWidth: 100, // Maximum preview image width
       imageHeight: 50, // Maximum preview image height
       renderCitations: true,